updated for version 7.3.089 Problem:    Compiler warning on 64 bit MS-Windows. Solution:   Add type cast. (Mike Williams)

+++ src/netbeans.c
@@ -800,7 +800,7 @@ netbeans_read()
 	 *		    -> gui event loop or select loop
 	 *			-> netbeans_read()
 	 */
-	save((char_u *)DETACH_MSG, strlen(DETACH_MSG));
+	save((char_u *)DETACH_MSG, (int)strlen(DETACH_MSG));
 	nb_close_socket();
 
 	if (len < 0)
